Fix: fixed the details product screen

+++ app/src/main/java/com/paway/mobileapplication/inventory/data/remote/InvoiceItemDto.kt
@@ -7,7 +7,7 @@ data class InvoiceItemDto(
     @SerializedName("id") val id: String,
     @SerializedName("description") val description: String,
     @SerializedName("quantity") val quantity: Int,
-    @SerializedName("unitPrice") val unitPrice: Int,
+    @SerializedName("unitPrice") val unitPrice: Double,
     @SerializedName("productId") val productId: String
 )
 fun InvoiceItemDto.toItem()= InvoiceItem( id,quantity)
